rename files using hyphens
