+++ packages/boilerplate-generator/template-web.browser.js
