+++ packages/boilerplate-generator/template-web.cordova.js
